requirements.txt
requests==2.31.0
anthropic==0.25.0
tweepy==4.14.0
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.0
apscheduler==3.10.4
jinja2==3.1.3
